refactor: Presentation 레이어 테스트 공통 설정 분리
