+++ src/test/java/com/example/realworld/application/ControllerTest.java
@@ -1,0 +1,43 @@
+package com.example.realworld.application;
+
+import com.example.realworld.application.users.persistence.repository.UserRepository;
+import com.example.realworld.application.users.service.UserService;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.mock.web.MockHttpSession;
+import org.springframework.test.web.servlet.MockMvc;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+public abstract class ControllerTest {
+
+    @Autowired
+    protected MockMvc mockMvc;
+
+    @Autowired
+    protected ObjectMapper mapper;
+
+    protected MockHttpSession session;
+
+    @Autowired
+    protected UserService userService;
+
+    @Autowired
+    protected UserRepository userRepository;
+
+    @BeforeEach
+    void setUp() {
+        session = new MockHttpSession();
+        session.setAttribute("email", "seokrae@gmail.com");
+    }
+
+    @AfterEach
+    void tearDown() {
+        userRepository.deleteAll();
+    }
+
+}
